Update student component to StudentChat
